+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 All notable changes to this project are documented here.
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+- `BootstrapSummary.round` parameter `precision` now defaults to `Literal["auto"]`.
 
 ## [0.1.1] - 2025-11-03